backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
email-validator>=2.2.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.19
pyjwt>=2.9.0
bcrypt>=4.1.0
scikit-learn>=1.5.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
httpx>=0.27.0
openai>=1.57.0
